add game class to gameplay.js, combine keyboard with game-screen
