+++ gameplay/gameplay.css
@@ -1,5 +1,7 @@
-.word{
-    display: flex;
-    justify-content: center;
-    font-size: 50px;
+.word {
+  position: fixed;
+  top: 20px;
+  display: flex;
+  justify-content: center;
+  font-size: 50px;
 }
