+++ gameplay/gameplay.js
@@ -1,3 +1,13 @@
+import { Game } from "./game/game.js"
+
+// Create Game 
+const gameScreen = document.getElementById("game_screen")
+
+const game = new Game(gameScreen, 1, 'normal', 'python')
+
+
+
+
 // Variables for word 
 let pickedLanguages = []
 let letterToTypeIndex = 0
@@ -7,13 +17,13 @@ const wordLetters = document.getElementById("word-container")
 
 
 // Import words.json
-async function importWords(){
-    try{
-       const response = await fetch("./gameplay/words.json")
+async function importWords() {
+    try {
+        const response = await fetch("./gameplay/words.json")
         const data = await response.json()
         let wordList = data.words
         return wordList
-    }catch(error){
+    } catch (error) {
         console.error(error)
         throw error
     }
@@ -24,17 +34,17 @@ const wordList = await importWords()
 
 // Check local storage
 
-if(!localStorage.getItem("pickedLanguages")){
+if (!localStorage.getItem("pickedLanguages")) {
     // TEMPORARY CODE - go back to mission configure page to pick languages
     window.location.href = "http://127.0.0.1:5502/missionConfigure.html"
-}else {
+} else {
     pickedLanguages = JSON.parse(localStorage.getItem("pickedLanguages"))
     generateWord()
 }
 
 // Display word function
-function displayWord(word){
-    for(let char of word){
+function displayWord(word) {
+    for (let char of word) {
         const span = document.createElement("span")
         const node = document.createTextNode(char)
         span.appendChild(node)
@@ -49,7 +59,7 @@ function clearWordDisplay() {
 }
 
 // Generate a new word
-function generateWord(){
+function generateWord() {
     // Get random language from picked language
     let randomIndex = Math.floor(Math.random() * pickedLanguages.length)
     let randomLanguage = pickedLanguages[randomIndex]
@@ -65,28 +75,45 @@ function generateWord(){
 
 
 // Update letter display, if correct letter typed, change color
-function updateDisplay(key){
+function attack(key) {
     let letterToType = wordLetters.getElementsByTagName("span")[letterToTypeIndex]
-    if (key === letterToType.innerHTML){
-            letterToType.style.color = "green"
-            letterToTypeIndex += 1
-        }
-
+    if (key === letterToType.innerHTML) {
+        letterToType.style.color = "green"
+        letterToTypeIndex += 1
+        game.onPlayerAttack()
+        console.log(`Attack ${key}`)
+    } else {
+        game.onPlayerAttack(false)
+        onsole.log('Missed')
+    }
     checkForCompletion(letterToTypeIndex)
 }
 
 
 // Check if word is fully typed, if it is then clear word and generate new word
-function checkForCompletion(letterToTypeIndex){
-    if (letterToTypeIndex === displayedWord.length){
+function checkForCompletion(letterToTypeIndex) {
+    if (letterToTypeIndex === displayedWord.length) {
         clearWordDisplay()
         generateWord()
-    } 
+    }
 }
 
 
 // EVENT LISTENER ON PAGE KEYDOWN
 document.addEventListener("keydown", event => {
     event.preventDefault()
-    updateDisplay(event.key)
+    attack(event.key)
+})
+
+
+
+// Game buttons
+const startBtn = document.querySelector('#start')
+const pauseBtn = document.querySelector('#pause')
+
+// start game
+startBtn.addEventListener('click', () => {
+    game.start()
+    startBtn.style.display = 'none'
+    pauseBtn.style.display = 'block'
 })
